+++ modules/cirrus/feeder/data.tf
@@ -1,0 +1,3 @@
+locals {
+  name_main = lower("${var.resource_prefix}-feeder-${var.feeder_config.name}")
+}